DLAPI-1000: add service accounts
