+++ src/components/middlewares/app-auth.ts
@@ -2,7 +2,7 @@ import {NextFunction, Request, Response} from '@gravity-ui/expresskit';
 
 import {AUTHORIZATION_HEADER, AUTHORIZATION_HEADER_VALUE_PREFIX} from '../../constants/header';
 import {decodeId} from '../../utils/ids';
-import {JwtAuth} from '../jwt-auth';
+import {verifyAccessToken} from '../jwt-auth';
 
 export const appAuth = async (req: Request, res: Response, next: NextFunction) => {
     req.ctx.log('AUTH');
@@ -16,7 +16,7 @@ export const appAuth = async (req: Request, res: Response, next: NextFunction) =
             try {
                 req.ctx.log('CHECK_ACCESS_TOKEN');
 
-                const payload = JwtAuth.verifyAccessToken({ctx: req.ctx, accessToken});
+                const payload = verifyAccessToken({ctx: req.ctx, accessToken});
 
                 if (payload.type === 'service_account') {
                     req.originalContext.set('subject', {
