+++ src/controllers/auth/logout.ts
@@ -2,7 +2,7 @@ import {AppRouteHandler, Response} from '@gravity-ui/expresskit';
 
 import {ApiTag} from '../../components/api-docs';
 import {clearAuthCookies, getAuthCookies} from '../../components/cookies';
-import {JwtAuth} from '../../components/jwt-auth';
+import {closeSession} from '../../components/jwt-auth';
 import {CONTENT_TYPE_JSON} from '../../constants/content-type';
 import {SuccessResponseModel, cookieHeaderSchema, successModel} from '../reponse-models';
 
@@ -13,7 +13,7 @@ export const logoutController: AppRouteHandler = async (
     const {authCookie} = getAuthCookies(req);
 
     if (authCookie && authCookie.refreshToken) {
-        await JwtAuth.closeSession({ctx: req.ctx}, {refreshToken: authCookie.refreshToken});
+        await closeSession({ctx: req.ctx}, {refreshToken: authCookie.refreshToken});
     }
 
     clearAuthCookies(req, res);
